backend/requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.6
Flask-Cors==5.0.1
